Added a default command timeout.
Previosuly, a command timeout was only used for commands that we expect to timeout, like forkbomb.
However, some slow, but possibly correct, vm implementations are hitting the prompt timeout, which
is in walltime. Using the command timeout allows us to specify timeouts in simtime (good) and
control things on a finer granularity. The current default command timeout is 300s, which may
be quite a bit too large.