Added usage stat tracking.
When test161 (client) runs a target, test group, or tag, the results are
logged in .test161/usage/. When users submit a target, a separate instance
of test161 runs that handles uploading collected usage files to the
test161-server. The server simply process the files and dumps the data into the
MongoDB instance.