Added support for metatargets.
Metatargets consist of two or more subtargets that together make up the assignment.
Running or submitting the metatarget will run/submit all subtargets, whereas
running or submitting a subtarget will run/submit the assignment up to that point.

On the client, we break down scores by subtarget and metatarget for both running and
submission.

On the server, we split the submission into multiple submissions and group/display
them accordingly.